feat(mlop-2760): creating metadata module
