--- butterfree/pipelines/feature_set_pipeline_metadata.py
@@ -1,228 +1,0 @@
-from __future__ import annotations
-
-from typing import TYPE_CHECKING, List, Literal, Optional, Union
-
-from pydantic import BaseModel, ConfigDict, Field, field_serializer
-from typing_extensions import Annotated
-
-from butterfree.extract.readers.reader_metadata import (
-    FileReaderMetadata,
-    KafkaReaderMetadata,
-    TableReaderMetadata,
-)
-from butterfree.load.writers.writer import WriterMetadata
-
-if TYPE_CHECKING:
-    from butterfree.pipelines import FeatureSetPipeline
-
-
-class FeatureMetadata(BaseModel):
-    """Metadata model for a column in a feature set.
-
-    This model represents the metadata of a single column in a feature set,
-    including its name, data type, and whether it's a primary key.
-    """
-
-    model_config = ConfigDict(arbitrary_types_allowed=True)
-
-    name: str = Field(..., description="The name of the column")
-    data_type: str = Field(
-        ...,
-        description=(
-            "The data type of the column (e.g., StringType, IntegerType) represented by pyspark.sql.types"  # noqa: E501
-        ),
-    )
-    description: str = Field(
-        ...,
-        description="The description of the column",
-    )
-    primary_key: bool = Field(
-        ...,
-        description="Whether the column is a primary (or partition if it's a Cassandra table) key",  # noqa: E501
-    )
-
-
-class FeatureSetMetadata(BaseModel):
-    """Metadata model for a feature set catalog.
-
-    This model represents the catalog information of a feature set,
-    including its name, description, and column definitions.
-    """
-
-    feature_set_name: str = Field(..., description="The name of the Feature Set")
-    description: str = Field(..., description="The description of the Feature Set")
-    columns: List[FeatureMetadata] = Field(
-        ..., description="A list of column definitions"
-    )
-
-
-class FeatureSetPipelineMetadata(BaseModel):
-    """Metadata model for a feature set pipeline.
-
-    This model represents the complete metadata of a feature set pipeline,
-    including its configuration, data sources, output schema, and processing details.
-    """
-
-    # To accept arbitrary types like FeatureSetPipeline
-    model_config = ConfigDict(arbitrary_types_allowed=True)
-
-    feature_set_pipeline: FeatureSetPipeline = Field(
-        ..., description="The feature set pipeline to create metadata from"
-    )
-    entity: str = Field(
-        ..., description="The entity type associated with the feature set"
-    )
-    feature_set_type: Literal["FeatureSet", "AggregatedFeatureSet"] = Field(
-        ..., description="The type of feature set"
-    )
-    windows_definition: Optional[List[str]] = Field(
-        None, description="The definition of the windows for the feature set"
-    )
-
-    # Required for correct serialization using Union
-    readers: List[
-        Annotated[
-            Union[FileReaderMetadata, KafkaReaderMetadata, TableReaderMetadata],
-            Field(discriminator="type"),
-        ]
-    ] = Field(
-        ...,
-        description="A list of data sources required to generate the feature set",
-    )
-
-    writers: List[WriterMetadata] = Field(
-        ..., description="The writers to be used for the feature set"
-    )
-    catalog: FeatureSetMetadata = Field(
-        ..., description="Metadata about the feature set's output"
-    )
-
-    @field_serializer("feature_set_pipeline")
-    def serialize_feature_set_pipeline(self, pipeline: FeatureSetPipeline) -> str:
-        """Serialize feature_set_pipeline to its class name."""
-        return pipeline.__class__.__name__
-
-    @classmethod
-    def _get_windows_definition(
-        cls, feature_set_pipeline: FeatureSetPipeline
-    ) -> Optional[List[str]]:
-        """Get windows definition if the feature set is an AggregatedFeatureSet.
-
-        Args:
-            feature_set_pipeline: The FeatureSetPipeline to get windows from.
-
-        Returns:
-            Optional list of window definitions or None if not applicable.
-        """
-        if (
-            isinstance(feature_set_pipeline.feature_set, AggregatedFeatureSet)
-            and feature_set_pipeline.feature_set._windows
-        ):
-            return [
-                window.frame_boundaries.window_definition
-                for window in feature_set_pipeline.feature_set._windows
-            ]
-        return None
-
-    @classmethod
-    def _create_catalog(
-        cls, feature_set_pipeline: FeatureSetPipeline
-    ) -> FeatureSetMetadata:
-        """Create a Catalog object from the feature set pipeline.
-
-        Args:
-            feature_set_pipeline: The FeatureSetPipeline to create catalog from.
-
-        Returns:
-            A Catalog instance with feature set metadata.
-        """
-
-        # This method already deals with the colums name transformation
-        catalog_schema = feature_set_pipeline.feature_set.get_schema()
-
-        columns = [
-            FeatureMetadata(
-                name=column["column_name"],
-                data_type=column[
-                    "type"
-                ].__class__.__name__,  # The get_schema method returns pyspark.sql.types
-                description=column["description"],
-                primary_key=column["primary_key"],
-            )
-            for column in catalog_schema
-        ]
-
-        return FeatureSetMetadata(
-            feature_set_name=feature_set_pipeline.feature_set.name,
-            description=feature_set_pipeline.feature_set.description,
-            columns=columns,
-        )
-
-    @classmethod
-    def _is_incremental(cls, feature_set_pipeline: FeatureSetPipeline) -> bool:
-        has_incremental_reader = any(
-            [
-                reader.incremental_strategy
-                for reader in feature_set_pipeline.source.readers
-            ]
-        )
-
-        return has_incremental_reader
-
-    @classmethod
-    def _get_key_features(
-        cls, feature_set_pipeline: FeatureSetPipeline
-    ) -> List[FeatureMetadata]:
-        """Get the key features of the feature set.
-
-        Args:
-            feature_set_pipeline: The FeatureSetPipeline to get key features from.
-
-        Returns:
-            A list of Column instances representing the key features.
-        """
-        return [
-            FeatureMetadata(
-                name=key.name,
-                data_type=key.dtype,
-                primary_key=True,
-            )
-            for key in feature_set_pipeline.feature_set.keys
-        ]
-
-    @classmethod
-    def from_pipeline(
-        cls, feature_set_pipeline: FeatureSetPipeline
-    ) -> "FeatureSetPipelineMetadata":
-        """Create a FeatureSetPipelineMetadata from a FeatureSetPipeline.
-
-        Args:
-            feature_set_pipeline: The FeatureSetPipeline to create metadata from.
-
-        Returns:
-            A FeatureSetPipelineMetadata instance.
-        """
-        feature_set_type = type(feature_set_pipeline.feature_set).__name__
-        is_incremental = cls._is_incremental(feature_set_pipeline)
-        windows_definition = cls._get_windows_definition(feature_set_pipeline)
-        catalog = cls._create_catalog(feature_set_pipeline)
-
-        writers = [
-            writer.build_metadata(feature_set_pipeline=feature_set_pipeline)
-            for writer in feature_set_pipeline.sink.writers
-        ]
-
-        readers = [
-            reader.build_metadata() for reader in feature_set_pipeline.source.readers
-        ]
-
-        return cls(
-            feature_set_pipeline=feature_set_pipeline,
-            is_incremental=is_incremental,
-            readers=readers,
-            catalog=catalog,
-            entity=feature_set_pipeline.feature_set.entity,
-            writers=writers,
-            feature_set_type=feature_set_type,
-            windows_definition=windows_definition,
-        )
